Update 32.rust_web_full_stack
* Add db access

+++ 32.rust_web_full_stack/server/src/models.rs
@@ -4,8 +4,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct Course {
-  pub teacher_id: usize,
-  pub id: Option<usize>,
+  pub teacher_id: i32,
+  pub id: Option<i32>,
   pub name: String,
   pub time: Option<NaiveDateTime>,
 }
